feat: add new forlder requests
now the new requests folder stores functions of requests to the server.
Also rewritten the context for scores using react-query
