+++ src/App.tsx
@@ -34,8 +34,8 @@ import MainPage from './pages/MainPage/MainPage';
 import ModulesPage from './pages/ModulesPage/ModulesPage';
 import MyCreatedCoursesPage from './pages/MyCreatedCoursesPage/MyCreatedCoursesPage';
 import MyPurchasedCoursesPage from './pages/MyPurchasedCoursesPage/MyPurchasedCoursesPage';
-import { DeployEnum } from './types';
 import UserPage from './pages/UserPage/UserPage';
+import { DeployEnum } from './types';
 
 const isProduction = process.env.REACT_APP_ENVIRONMENT === 'production';
 const tg = window.Telegram.WebApp;
@@ -49,7 +49,7 @@ function App() {
   const eventBuilder = useTWAEvent();
   const { showNotification } = useNotification();
   const { initData } = retrieveLaunchParams();
-  const { refreshPoints, setPoints } = usePoints();
+  const { refreshPoints } = usePoints();
   const { setCourseContractBalance, setPurchaseContractBalance } =
     useContract();
 
@@ -83,7 +83,7 @@ function App() {
         showNotification(message, status);
 
         if (points) {
-          setPoints(points);
+          refreshPoints();
         }
 
         if (balance && type) {
@@ -106,7 +106,6 @@ function App() {
     showNotification,
     initData?.user?.id,
     refreshPoints,
-    setPoints,
     setCourseContractBalance,
     setPurchaseContractBalance,
     eventBuilder,
@@ -118,72 +117,51 @@ function App() {
   }, [initData?.user?.languageCode]);
 
   return (
-    <TonConnectUIProvider manifestUrl={manifestUrl}>
-      <I18nextProvider i18n={i18n}>
-        <QueryClientProvider client={queryClient}>
-          <Router>
-            <Routes>
-              {/* all users */}
-              <Route path="/" element={<MainPage />} />
-              <Route path="/user" element={<UserPage />} />
-              <Route path="/course/:courseId" element={<CourseDetailsPage />} />
-              <Route
-                path="course/category/:category"
-                element={<CoursesOneCategoryPage />}
-              />
-              <Route path="/course/create" element={<CreateCourseFormPage />} />
-              {/* seller */}
-              <Route
-                path="/course/edit/:courseId"
-                element={<EditCourseFormPage />}
-              />
-              <Route
-                path="/course-part/create/:type/:parentId"
-                element={<CreateCoursePartPage />}
-              />
-              <Route
-                path="/course-part/edit/:parentId/:type/:itemId"
-                element={<EditCoursePartPage />}
-              />
-              <Route
-                path="/course/created"
-                element={<MyCreatedCoursesPage />}
-              />
-              {/* customer */}
-              <Route
-                path="/course/purchased"
-                element={<MyPurchasedCoursesPage />}
-              />
-              {/* customer & seller */}
-              <Route
-                path="/module/course/:courseId"
-                element={<ModulesPage />}
-              />
-              <Route
-                path="/lesson/module/:moduleId"
-                element={<LessonsPage />}
-              />
-              <Route path="/lesson/:lessonId" element={<LessonPage />} />
-            </Routes>
-            <ToastContainer
-              position="top-center"
-              autoClose={5000}
-              hideProgressBar={false}
-              newestOnTop={false}
-              closeOnClick
-              rtl={false}
-              pauseOnFocusLoss
-              draggable
-              pauseOnHover
-              theme="colored"
-              transition={Bounce}
-            />
-            <ModalEarnPoints />
-          </Router>
-          {!isProduction && <ReactQueryDevtools />}
-        </QueryClientProvider>
-      </I18nextProvider>
-    </TonConnectUIProvider>
+    <Router>
+      <Routes>
+        {/* all users */}
+        <Route path="/" element={<MainPage />} />
+        <Route path="/user" element={<UserPage />} />
+        <Route path="/course/:courseId" element={<CourseDetailsPage />} />
+        <Route
+          path="course/category/:category"
+          element={<CoursesOneCategoryPage />}
+        />
+        <Route path="/course/create" element={<CreateCourseFormPage />} />
+        {/* seller */}
+        <Route path="/course/edit/:courseId" element={<EditCourseFormPage />} />
+        <Route
+          path="/course-part/create/:type/:parentId"
+          element={<CreateCoursePartPage />}
+        />
+        <Route
+          path="/course-part/edit/:parentId/:type/:itemId"
+          element={<EditCoursePartPage />}
+        />
+        <Route path="/course/created" element={<MyCreatedCoursesPage />} />
+        {/* customer */}
+        <Route path="/course/purchased" element={<MyPurchasedCoursesPage />} />
+        {/* customer & seller */}
+        <Route path="/module/course/:courseId" element={<ModulesPage />} />
+        <Route path="/lesson/module/:moduleId" element={<LessonsPage />} />
+        <Route path="/lesson/:lessonId" element={<LessonPage />} />
+      </Routes>
+      <ToastContainer
+        position="top-center"
+        autoClose={5000}
+        hideProgressBar={false}
+        newestOnTop={false}
+        closeOnClick
+        rtl={false}
+        pauseOnFocusLoss
+        draggable
+        pauseOnHover
+        theme="colored"
+        transition={Bounce}
+      />
+      <ModalEarnPoints />
+    </Router>
+    // </TonConnectUIProvider>
   );
 }
 
@@ -198,14 +176,21 @@ export default () => (
       process.env.REACT_APP_TELEMETREE_APPLICATION_NAME || 'courses_app_name'
     }
   >
-    <NotificationProvider>
+    <QueryClientProvider client={queryClient}>
       <PointsProvider>
-        <ContractProvider>
-          <ModalProvider>
-            <App />
-          </ModalProvider>
-        </ContractProvider>
+        <NotificationProvider>
+          <ContractProvider>
+            <ModalProvider>
+              <I18nextProvider i18n={i18n}>
+                <TonConnectUIProvider manifestUrl={manifestUrl}>
+                  <App />
+                </TonConnectUIProvider>
+              </I18nextProvider>
+            </ModalProvider>
+          </ContractProvider>
+        </NotificationProvider>
       </PointsProvider>
-    </NotificationProvider>
+      {!isProduction && <ReactQueryDevtools />}
+    </QueryClientProvider>
   </TwaAnalyticsProvider>
 );
