+++ src/functions/index.ts
@@ -1,19 +1,7 @@
+export { fetchPointsAPI } from '../requests/points';
 export { createAxiosWithAuth } from './axiosWithAuth';
 export { capitalizeFirstLetter } from './capitalizeFirstLetter';
-export {
-  createOrUpdateCourseAPI,
-  deleteCourseAPI,
-  fetchAllCoursesAPI,
-  fetchAllMyCreatedCourses,
-  fetchAllMyPurchasedCourses,
-  fetchCourseDetailsAPI,
-  groupCoursesByCategory,
-} from './course';
-export {
-  deleteCoursePartAPI,
-  fetchLessonsAPI,
-  fetchModulesAPI,
-} from './coursePart';
+export { groupCoursesByCategory } from './course';
 export { encodeSaleMessage } from './encodeSaleMessage';
 export { handleAuthError } from './errorHandlers';
 export { getCSSVariableValue } from './getCSSVariableValue';
