+++ src/pages/EditCourseFormPage/EditCourseFormPage.tsx
@@ -5,8 +5,8 @@ import { useParams } from 'react-router-dom';
 import CourseForm from '../../components/CourseForm/CourseForm';
 import Header from '../../components/Header/Header';
 import { CourseFormProvider } from '../../context';
-import { fetchCourseDetailsAPI } from '../../functions';
 import { useCourseForm } from '../../hooks';
+import { fetchCourseDetailsAPI } from '../../requests';
 import Container from '../../ui/Container/Container';
 import { Loader } from '../../ui/Loader/Loader';
 import { MessageBox } from '../../ui/MessageBox/MessageBox';
