+++ src/pages/MyCreatedCoursesPage/MyCreatedCoursesPage.tsx
@@ -4,8 +4,8 @@ import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import CourseItem from '../../components/CourseItem/CourseItem';
 import Points from '../../components/Points/Points';
-import { fetchAllMyCreatedCourses } from '../../functions';
 import { filterCourses } from '../../functions/filterCourses';
+import { fetchAllMyCreatedCourses } from '../../requests';
 import { ICourse } from '../../types';
 import Container from '../../ui/Container/Container';
 import { Loader } from '../../ui/Loader/Loader';
