+++ src/pages/MyPurchasedCoursesPage/MyPurchasedCoursesPage.tsx
@@ -4,8 +4,8 @@ import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import CourseItem from '../../components/CourseItem/CourseItem';
 import Points from '../../components/Points/Points';
-import { fetchAllMyPurchasedCourses } from '../../functions';
 import { filterCourses } from '../../functions/filterCourses';
+import { fetchAllMyPurchasedCourses } from '../../requests';
 import { ICourse } from '../../types';
 import Container from '../../ui/Container/Container';
 import { Loader } from '../../ui/Loader/Loader';
