+++ src/requests/index.ts
@@ -1,0 +1,14 @@
+export {
+  createOrUpdateCourseAPI,
+  deleteCourseAPI,
+  fetchAllCoursesAPI,
+  fetchAllMyCreatedCourses,
+  fetchAllMyPurchasedCourses,
+  fetchCourseDetailsAPI,
+} from './course';
+export {
+  deleteCoursePartAPI,
+  fetchLessonsAPI,
+  fetchModulesAPI,
+} from './coursePart';
+export { fetchPointsAPI } from './points';
